deleteChatController updated

+++ controllers/chatControllers/deleteChatController.js
@@ -4,18 +4,19 @@ const User = require('../../models/userModel')
 const Comment = require('../../models/commentModel')
 const { StatusCodes } = require('http-status-codes')
 const getCreatorIdFromHeaders = require('../../middlewares/getCreatorIdFromHeaders')
-const mongoose = require('mongoose')
 
 const deleteChat = tryCatchWrapper(async (req, res) => {
     const chatId = req.params.chatId
-    const ownerId = await getCreatorIdFromHeaders(req.headers)
-    const chat = await Chat.findOne({ _id: chatId, owner: ownerId })
-    console.log(chat)
+    const deleterUser = await getCreatorIdFromHeaders(req.headers)
+    const chat = await Chat.findOne({ _id: chatId })
     if (chat) {
-        await Chat.findByIdAndDelete(chatId)
-        await User.updateMany({ chats: { $in: [chatId] } }, { $pull: { chats: chatId } })
-        await Comment.deleteMany({ '_id.room_id': chatId })
-        return res.status(StatusCodes.OK).json({ message: 'Chat deleted successfully.' })
+        if (chat.creator_id == deleterUser && chat.is_private == true) {
+            await Chat.findByIdAndDelete(chatId)
+            await User.updateMany({ chats: { $in: [chatId] } }, { $pull: { chats: chatId } })
+            await Comment.deleteMany({ '_id.room_id': chatId })
+            return res.status(StatusCodes.OK).json({ message: 'Chat deleted successfully.' })
+        }
+        return res.status(StatusCodes.FORBIDDEN).json({ message: 'You are not allowed to delete this chat.' })
     }
     return res.status(StatusCodes.NOT_FOUND).json({ message: 'No chat found.' })
 })
